Use a context provider for global styles config
