+++ packages/edit-site/src/components/editor/index.js
@@ -41,7 +41,8 @@ import InserterSidebar from '../secondary-sidebar/inserter-sidebar';
 import ListViewSidebar from '../secondary-sidebar/list-view-sidebar';
 import ErrorBoundary from '../error-boundary';
 import { store as editSiteStore } from '../../store';
-import { useGlobalStylesRenderer } from './global-styles-renderer';
+import { GlobalStylesRenderer } from './global-styles-renderer';
+import { GlobalStylesProvider } from '../global-styles/global-styles-provider';
 
 const interfaceLabels = {
 	secondarySidebar: __( 'Block Library' ),
@@ -159,8 +160,6 @@ function Editor( { initialSettings, onError } ) {
 		}
 	}, [ isNavigationOpen ] );
 
-	useGlobalStylesRenderer();
-
 	// Don't render the Editor until the settings are set and loaded
 	if ( ! settings?.siteUrl ) {
 		return null;
@@ -186,87 +185,94 @@ function Editor( { initialSettings, onError } ) {
 						type={ templateType }
 						id={ entityId }
 					>
-						<BlockContextProvider value={ blockContext }>
-							<ErrorBoundary onError={ onError }>
-								<FullscreenMode isActive />
-								<UnsavedChangesWarning />
-								<KeyboardShortcuts.Register />
-								<SidebarComplementaryAreaFills />
-								<InterfaceSkeleton
-									labels={ interfaceLabels }
-									drawer={ <NavigationSidebar /> }
-									secondarySidebar={ secondarySidebar() }
-									sidebar={
-										sidebarIsOpened && (
-											<ComplementaryArea.Slot scope="core/edit-site" />
-										)
-									}
-									header={
-										<Header
-											openEntitiesSavedStates={
-												openEntitiesSavedStates
-											}
-										/>
-									}
-									notices={ <EditorSnackbars /> }
-									content={
-										<>
-											<EditorNotices />
-											{ template && (
-												<BlockEditor
-													setIsInserterOpen={
-														setIsInserterOpened
-													}
-												/>
-											) }
-											{ templateResolved &&
-												! template &&
-												settings?.siteUrl &&
-												entityId && (
-													<Notice
-														status="warning"
-														isDismissible={ false }
-													>
-														{ __(
-															"You attempted to edit an item that doesn't exist. Perhaps it was deleted?"
-														) }
-													</Notice>
-												) }
-											<KeyboardShortcuts />
-										</>
-									}
-									actions={
-										<>
-											{ isEntitiesSavedStatesOpen ? (
-												<EntitiesSavedStates
-													close={
-														closeEntitiesSavedStates
-													}
-												/>
-											) : (
-												<div className="edit-site-editor__toggle-save-panel">
-													<Button
-														variant="secondary"
-														className="edit-site-editor__toggle-save-panel-button"
-														onClick={
-															openEntitiesSavedStates
+						<GlobalStylesProvider>
+							<BlockContextProvider value={ blockContext }>
+								<GlobalStylesRenderer />
+								<ErrorBoundary onError={ onError }>
+									<FullscreenMode isActive />
+									<UnsavedChangesWarning />
+									<KeyboardShortcuts.Register />
+									<SidebarComplementaryAreaFills />
+									<InterfaceSkeleton
+										labels={ interfaceLabels }
+										drawer={ <NavigationSidebar /> }
+										secondarySidebar={ secondarySidebar() }
+										sidebar={
+											sidebarIsOpened && (
+												<ComplementaryArea.Slot scope="core/edit-site" />
+											)
+										}
+										header={
+											<Header
+												openEntitiesSavedStates={
+													openEntitiesSavedStates
+												}
+											/>
+										}
+										notices={ <EditorSnackbars /> }
+										content={
+											<>
+												<EditorNotices />
+												{ template && (
+													<BlockEditor
+														setIsInserterOpen={
+															setIsInserterOpened
 														}
-														aria-expanded={ false }
-													>
-														{ __(
-															'Open save panel'
-														) }
-													</Button>
-												</div>
-											) }
-										</>
-									}
-									footer={ <BlockBreadcrumb /> }
-								/>
-								<Popover.Slot />
-								<PluginArea />
-							</ErrorBoundary>
-						</BlockContextProvider>
+													/>
+												) }
+												{ templateResolved &&
+													! template &&
+													settings?.siteUrl &&
+													entityId && (
+														<Notice
+															status="warning"
+															isDismissible={
+																false
+															}
+														>
+															{ __(
+																"You attempted to edit an item that doesn't exist. Perhaps it was deleted?"
+															) }
+														</Notice>
+													) }
+												<KeyboardShortcuts />
+											</>
+										}
+										actions={
+											<>
+												{ isEntitiesSavedStatesOpen ? (
+													<EntitiesSavedStates
+														close={
+															closeEntitiesSavedStates
+														}
+													/>
+												) : (
+													<div className="edit-site-editor__toggle-save-panel">
+														<Button
+															variant="secondary"
+															className="edit-site-editor__toggle-save-panel-button"
+															onClick={
+																openEntitiesSavedStates
+															}
+															aria-expanded={
+																false
+															}
+														>
+															{ __(
+																'Open save panel'
+															) }
+														</Button>
+													</div>
+												) }
+											</>
+										}
+										footer={ <BlockBreadcrumb /> }
+									/>
+									<Popover.Slot />
+									<PluginArea />
+								</ErrorBoundary>
+							</BlockContextProvider>
+						</GlobalStylesProvider>
 					</EntityProvider>
 				</EntityProvider>
 			</SlotFillProvider>
