+++ packages/edit-site/src/components/global-styles/context.js
@@ -1,0 +1,15 @@
+/**
+ * WordPress dependencies
+ */
+import { createContext } from '@wordpress/element';
+
+export const DEFAULT_GLOBAL_STYLES_CONTEXT = {
+	user: {},
+	base: {},
+	merged: {},
+	setUserConfig: () => {},
+};
+
+export const GlobalStylesContext = createContext(
+	DEFAULT_GLOBAL_STYLES_CONTEXT
+);
